requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
ultralytics==8.0.220
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
pyyaml==6.0.1
pillow==10.1.0
matplotlib==3.8.2
scikit-learn==1.3.2
pandas==2.1.4
seaborn==0.13.0
tqdm==4.66.1